feat: added wallet service

+++ app/internal/service/wallet.go
@@ -2,19 +2,21 @@ package service
 
 import (
 	"context"
+	"time"
 
-	logger2 "github.com/mustan989/wallet/pkg/logger"
+	"github.com/mustan989/wallet/pkg/helper"
+	"github.com/mustan989/wallet/pkg/logger"
 	"github.com/mustan989/wallet/repository"
 	"github.com/mustan989/wallet/service"
 )
 
 type WalletOption func(w *wallet)
 
-func WithLogger(log logger2.Logger) WalletOption { return func(w *wallet) { w.log = log } }
+func WithLogger(log logger.Logger) WalletOption { return func(w *wallet) { w.log = log } }
 
 func NewWallet(repo repository.Wallet, options ...WalletOption) service.Wallet {
 	w := &wallet{
-		log:  logger2.Default(),
+		log:  logger.Default(),
 		repo: repo,
 	}
 
@@ -26,7 +28,7 @@ func NewWallet(repo repository.Wallet, options ...WalletOption) service.Wallet {
 }
 
 type wallet struct {
-	log logger2.Logger
+	log logger.Logger
 
 	repo repository.Wallet
 }
@@ -84,9 +86,24 @@ func (w *wallet) Update(ctx context.Context, request *service.WalletUpdateReques
 }
 
 func (w *wallet) DeleteByID(ctx context.Context, request *service.WalletDeleteByIDRequest) (*service.WalletDeleteByIDResponse, error) {
+	data, err := w.repo.FindByID(ctx, request.ID)
+	if err != nil {
+		return nil, err
+	}
+
+	if data.DeletedAt == nil {
+		data.DeletedAt = helper.Timep(time.Now())
+		if err = w.repo.Update(ctx, data); err != nil {
+			w.log.Errorf("Error marking wallet as deleted: %s", err)
+			return nil, err
+		}
+		return &service.WalletDeleteByIDResponse{Data: data}, nil
+	}
+
 	deleted, err := w.repo.DeleteByID(ctx, request.ID)
 	if err != nil {
 		w.log.Errorf("Error deleting wallet: %s", err)
+		return nil, err
 	}
 	return &service.WalletDeleteByIDResponse{Data: deleted}, nil
 }
